requirements.txt
anndata==0.10.6
array_api_compat==1.6
blinker==1.8.2
cachelib==0.13.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
distlib @ file:///private/tmp/python-distlib-20231212-4643-rs59kw/distlib-0.3.8
filelock @ file:///private/tmp/python-filelock-20231112-5255-1qu5zmd/filelock-3.13.1/dist/filelock-3.13.1-py3-none-any.whl#sha256=57dbda9b35157b05fb3e58ee91448612eb674172fab98ee235ccb0b5bee19a1c
Flask==3.0.3
Flask-Session==0.8.0
fonttools==4.50.0
h5py==3.10.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.3.2
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.3
msgspec==0.18.6
natsort==8.4.0
numpy==1.26.4
packaging==24.0
pandas==2.2.1
pillow==10.3.0
platformdirs @ file:///private/tmp/python-platformdirs-20231204-5092-dluipp/platformdirs-4.1.0
pydeseq2==0.4.8
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML @ file:///private/tmp/pyyaml-20231018-5353-1w5l4nf/PyYAML-6.0.1
redis==5.0.7
requests==2.32.3
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
spotipy==2.24.0
threadpoolctl==3.4.0
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.3
